Show some Cast Love, Alerts, Checks, Logging, Whoo!
I *think* I actually devised a way to make cast fetching non-blocking.
It's easy to do for 'nix systems, but for Windows, the user needs to
input the path to the PHP executable.

Settings now has an input field for the PHP executable, plus a bunch of
bells and whistles to check and display that info correctly.  Providing
the path isn't required, but it'll make cast fetching faster.

Add an alert box for various info/warning messages.

Add file checking, enabled module checking for PHP.  Need to flesh out
the list of required PHP modules...need to read throught he forum.

More tweaking on UI.