Okay Google, Say Hi to Alexa!
Holy shit...it supports Alexa now.
Or, it should.  I can't test the account linking, but I think all the
parts are there.  I need somebody to try it and tell me if it works or
not.
If it does work, it still needs a lot of love...but the basics should
work!