Minor tweaks
Add force-logout when the server detects invalid API token.
Use apiToken as session ID instead of plexServerToken, as this changes.
Fix issue saving client data.
Update parsePlayCommand to be a little smarter when asked to play stuff on a specific player.  Will need to make this multi-lingual, as "on the" shouldn't be hardcoded.
Fix JS null value error...