Only scan for playback devices once previous request finishes
I noticed when scanning for playback devices, phlex had a hard-coded 5 second delay before performing the next check - regardless of whether the previous request completed or not. This commit changes the device scan to only allow for a single request to run at a time (aborting the previous request if it is still pending), then sets a timeout (defined as a constant) if `autoDeviceScan` is set to true - which it is right now.

I noticed my plexkodiconnect instances popping up the disappearing, checked out my network panel and noticed a lot of those requests would not be completed by the time the next one fired. I'm hoping this will help with that!