Fallback requests, Fixes
Re-implement the hook to trigger fallback responses if Flex TV doesn't get back to the mothership in time.

Fix issues where sending a null string to compareTitles would throw an error.
Fix a warning in the log if open_basedir warnings are thrown trying to find php logs to read.

Add a param to continue a session for fallbacks...