Make sure return images are publicly accessible and served over HTTPS
Try to prefer direct/https connections for server, but also make sure every backup address is tested.

MultiCurl now has a "test" function, which, when passed an array of URL's, will return the *first* URL to return successful and stop execution - versus past behavior where it would CURL all of the URL's, wait for all responses, and then return whatever address was first.

Because we're not always referring to the server by it's public address, re-add the routine where we cache images, serve them up from Flex TV, and if necessary, pass them to a https image proxy.