Improve Device Title Matching, Duplicate Media Item Removal
Fix an issue tht would occur in requests to play media on a specific device where the media had an instance of the delimiter twice in the request string.

Fix an issue in removing duplicate media from playback requests when media exists on two shared servers, but neither of those servers is preferred.