Fun With CA Certs!
Wrap cert logic for CURL activities in a getCert() function.
If system CA certs are present/maintained, use those.
If system CA certs are not present, attempt to download to /rw folder.
As a last resort, use the one bundled with the project.
Remove extra copies of "project" CA Cert file.