Holy Toledo!!
Icons and colors and bears...oh my.
Add downloader sections. This can probably just be one card...but it's a start.
Auto-generate settings sections. Neato! It seems to all work.
Add a modal card if not in the commands window to display a result.
Fix console errors in recently_added because file's now inclded in the main PHP...